Add validation
We should guard that number arguments are within the range of the base segment
of the particular fraction tree. That is, that Farey tree numbers are within 0
and 1, octave reduced tree numbers are within 1 and 2, etc. Not doing so causes
run-away processes trying to find nodes for numbers outside of their tree's
range.

This commit adds a validate call to check a number is within the tree's base
segment, before executing some methods, like path_to.