Move to an instance implementation
Using classes to denote the types of fraction trees doesn't scale well. It
requires the user to know which class associates with which range of the tree,
and it doesn't support custom tree ranges easily.

All the former ability can be achieved using a class instance, and keywords to
denote the different known tree ranges. With one class, we can also expand the
constructor to accept keyword arguments, which lets us define custom tree
ranges more easily.

This change does the following:

1. Moves the behaviour of the various FractionTree sub-classes into the
   FractionTree class. This change also moves most formerly class methods to
   instance methods in the FractionTree class.
2. Fixes a bug in which a number whose value matches an endpoint, was causing
   #path_to to run-away. The method now includes a guard to return the endpoint
   when the value of number matches it.
3. Adds the #to_node method to the String class, to facilitate converting
   strings like "1/0" into FractionTree:Nodes.